make classroomRegDoc interfaces available
